+++ Backend/webapiproj/webapiproj/Models/Ram.cs
@@ -11,7 +11,7 @@ namespace webapiproj.Models
         public int Id { get; set; }
         public string Nev { get; set; }
         public string MemoriaTipus { get; set; }
-        public double Frekvencia { get; set; }
+        public int Frekvencia { get; set; }
         public int Meret { get; set; }
         public string KepNev { get; set; }
     }
